convert blit to cpp
